Adds Documentation and minor quality of life fixes

+++ openfold/utils/superimposition.py
@@ -35,8 +35,8 @@ def _superimpose_np(reference, coords):
 
 
 def _superimpose_single(reference, coords):
-    reference_np = reference.detach().cpu().numpy()    
-    coords_np = coords.detach().cpu().numpy()
+    reference_np = reference.detach().to(torch.float).cpu().numpy()    
+    coords_np = coords.detach().to(torch.float).cpu().numpy()
     superimposed, rmsd = _superimpose_np(reference_np, coords_np)
     return coords.new_tensor(superimposed), coords.new_tensor(rmsd)
